+++ R/EIvsPoF.R
@@ -192,7 +192,7 @@ optim.EIvsPoF = function(
     AF_Pareto = nsga2(
       fn=AF_EIvsPoF, idim=dim, odim=2,
       fgpi=fgpi, fmean=fmean, fsd=fsd, Cgpi=Cgpi, fmin=m2,
-      generations=100, popsize=100*dim,
+      generations=100, popsize=100*nprl,
       cprob=0.9, cdist=20, mprob=0.1, mdist=20,
       lower.bounds=rep(0, dim), upper.bounds=rep(1, dim))
     AF_PF = AF_Pareto$value # Pareto front
@@ -203,7 +203,8 @@ optim.EIvsPoF = function(
     # Here, a small disturbance is added to PF to avoid PS aggregation, 
     # resulting in clustering failure.
     AF_PF_dis = AF_PF + rnorm(length(AF_PF), mean = 0, sd = 0.001)
-    AF_cl = kmeans(x = AF_PF_dis, centers = nprl, nstart = 25)
+    AF_cl = kmeans(x = AF_PF_dis, centers = nprl, 
+                   iter.max = 100, nstart = 5*nprl)
     for (cl in 1:nprl) {
       ## a single member with highest EIC is selected from each cluster
       cl_idx = which(AF_cl$cluster == cl)