+++ R/PEPBO.R
@@ -10,7 +10,6 @@
 #' of the \code{matrix} determines the input dimension (\code{length(x)} in \code{blackbox(x)}); 
 #' the first column gives lower bounds and the second gives upper bounds
 #' @param nprl positive integer giving the number of the parallel points per iteration
-#' @param kappa positive weighted parameter in the LCB acquistion function
 #' @param equal an optional vector containing zeros and ones, whose length equals the number of
 #' constraints, specifying which should be treated as equality constraints (\code{1}) and 
 #' which as inequality (\code{0}) 
@@ -159,7 +158,7 @@
 #' EPBO$xbest
 
 optim.PEP = function(
-    blackbox, B, nprl=3, kappa = 1,
+    blackbox, B, nprl=3, 
     equal=FALSE, ethresh=1e-2, 
     Xstart=NULL, start=10, end=100, urate=ceiling(10/nprl), rho=NULL, 
     dg_start=c(1e-2*sqrt(nrow(B)), 1e-6), 
@@ -326,14 +325,15 @@ optim.PEP = function(
       fn=AF_EYvsSDY, idim=dim, odim=2,
       fgpi=fgpi, fmean=fmean, fsd=fsd, Cgpi=Cgpi,
       rho=rho, equal=equal,
-      generations=100, popsize=100*dim,
+      generations=100, popsize=100*nprl,
       cprob=0.9, cdist=20, mprob=0.1, mdist=20,
       lower.bounds=rep(0, dim), upper.bounds=rep(1, dim))
     AF_PF = AF_Pareto$value # Pareto front (-logSDY, EY)
     AF_PS = AF_Pareto$par   # Pareto set
+    PF_selected = matrix(NA, nrow = nprl, ncol = 2)
     
-    ## Remove Pareto sets with probability of feasibility less than 0.2
-    PS_PoF_idx = which(AF_PoF(AF_PS, Cgpi) > 0.2)
+    ## Remove Pareto sets with probability of feasibility less than 0.5
+    PS_PoF_idx = which(AF_PoF(AF_PS, Cgpi) > 0.5)
     if(length(PS_PoF_idx) > 0.2 * nrow(AF_PS)){
       AF_PF = AF_PF[PS_PoF_idx,]
       AF_PS = AF_PS[PS_PoF_idx,]
@@ -341,6 +341,7 @@ optim.PEP = function(
     
     ## pure exploitation (minimize the predictive mean)
     min_EY_idx = which.min(AF_PF[,2])
+    PF_selected[1,] = AF_PF[min_EY_idx, ]
     out_EY = optim( # enhance the EY approach
       par=AF_PS[min_EY_idx, ], fn=AF_EY, method="L-BFGS-B",
       lower=0, upper=1,
@@ -354,45 +355,33 @@ optim.PEP = function(
     xnext = unnormalize(xnext_unit, B)
     X = rbind(X, xnext)
     
-    PF_selected = matrix(NA, nrow = nprl, ncol = 2)
-    PF_selected[1,] = AF_PF[min_EY_idx, ]
-    
-    # AF_PF = AF_PF[-min_EY_idx,] 
-    # AF_PS = AF_PS[-min_EY_idx,] 
-    
     ## Perform k-means clustering on the AF's Pareto front
-    AF_cl = kmeans(x = AF_PF, centers = nprl-1, nstart = 25)
-    
-    # kappa = 1 + pnorm(6 * (k - start) / (end - start) - 3)
-    # kappa = 0.2 * dim * log(2*k) # https://github.com/kirthevasank/add-gp-bandits/blob/master/BOLibkky/getUCBUtility.m
-    # if(verb > 0) { cat("  kappa=", kappa, " ", sep="", "\n")}
-    
+    AF_PF_Scaled = min_max_scale(AF_PF)
+    AF_cl = kmeans(x = AF_PF_Scaled, centers = nprl-1, 
+                   iter.max = 100, nstart = 5*nprl)
+
     for (cl in 1:(nprl-1)) {
       ## a single member with highest EIC is selected from each cluster
       cl_idx = which(AF_cl$cluster == cl)
-      cl_PS = matrix(AF_PS[cl_idx, ], nrow = length(cl_idx))
-      cl_PF = matrix(AF_PF[cl_idx, ], nrow = length(cl_idx))
+      cl_PS = AF_PS[cl_idx, , drop=FALSE]
+      cl_PF = AF_PF[cl_idx, , drop=FALSE]
       
-      # ## LCB criterion
-      # cl_LCB = cl_PF[,2] - sqrt(kappa) * exp(-cl_PF[,1])
-      # min_LCB_idx = which.min(cl_LCB)
-      # PF_selected[cl+1,] = cl_PF[min_LCB_idx,]
       
-      ## center of the Pareto front
-      cl_center = AF_cl$centers[cl,]
-      dist = apply(cl_PF, 1, \(x){norm(x- cl_center, type="2")}) # Euclidean distance
-      min_dist_idx = which.min(dist)  # Find index of minimum distance
-      PF_selected[cl+1,] = cl_PF[min_dist_idx,]
+      # ## center of the Pareto front
+      # cl_center = AF_cl$centers[cl,]
+      # dist = apply(AF_PF_Scaled[cl_idx, , drop=FALSE], 1, 
+      #              function(x){norm(x- cl_center, type="2")}) # Euclidean distance
+      # min_dist_idx = which.min(dist)  # Find index of minimum distance
+      # PF_selected[cl+1,] = cl_PF[min_dist_idx,]
       
-      # ## max std criterion
-      # max_STD_idx = which.max(-cl_PF[,1])
-      # PF_selected[cl+1,] = cl_PF[max_STD_idx,]
+      ## max std criterion
+      max_STD_idx = which.max(-cl_PF[,1])
+      PF_selected[cl+1,] = cl_PF[max_STD_idx,]
       
       
       # calculate next point
-      # xnext_unit = cl_PS[min_LCB_idx, ]
-      xnext_unit = cl_PS[min_dist_idx, ]
-      # xnext_unit = cl_PS[max_STD_idx, ]
+      # xnext_unit = cl_PS[min_dist_idx, ]
+      xnext_unit = cl_PS[max_STD_idx, ]
       X_unit = rbind(X_unit, xnext_unit)
       xnext = unnormalize(xnext_unit, B)
       X = rbind(X, xnext)